requirements.txt
# ── Core Web Framework ──────────────────────────────────────────
fastapi==0.110.1
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Streamlit Frontend ───────────────────────────────────────────
streamlit==1.33.0
requests==2.31.0

# ── LLM ─────────────────────────────────────────────────────────
google-generativeai==0.5.4

# ── Agent Frameworks ─────────────────────────────────────────────
langchain==0.1.20
langchain-google-genai==1.0.3

# ── Database ─────────────────────────────────────────────────────
sqlalchemy==2.0.29
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# ── Vector Database ──────────────────────────────────────────────
chromadb==0.4.24
sentence-transformers==2.7.0

# ── Utilities ────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.0
pydantic-settings==2.2.1
httpx==0.27.0

# ── Testing ──────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
